feat: outsource initial layoutcolumn preparation
