+++ src/components/App/ExtensibilityRoutes.js
@@ -1,8 +1,8 @@
-import React, { Suspense, useEffect } from 'react';
+import React, { Suspense } from 'react';
 import pluralize from 'pluralize';
 import i18next from 'i18next';
-import { Route, useParams, useSearchParams } from 'react-router-dom';
-import { useRecoilState } from 'recoil';
+import { Route, useParams } from 'react-router-dom';
+import { useRecoilValue } from 'recoil';
 import { FlexibleColumnLayout } from '@ui5/webcomponents-react';
 import { useTranslation } from 'react-i18next';
 
@@ -13,6 +13,7 @@ import { usePrepareCreateProps } from 'resources/helpers';
 import { Spinner } from 'shared/components/Spinner/Spinner';
 import { ResourceCreate } from 'shared/components/ResourceCreate/ResourceCreate';
 import { ErrorBoundary } from 'shared/components/ErrorBoundary/ErrorBoundary';
+import { usePrepareLayoutColumns } from 'shared/hooks/usePrepareLayout';
 
 const List = React.lazy(() => import('../Extensibility/ExtensibilityList'));
 const Details = React.lazy(() =>
@@ -26,49 +27,20 @@ const ColumnWrapper = ({
   extension,
   urlPath,
 }) => {
-  const [layoutState, setLayoutColumn] = useRecoilState(columnLayoutState);
-  const [searchParams] = useSearchParams();
-  const layout = searchParams.get('layout');
+  const layoutState = useRecoilValue(columnLayoutState);
   const { resourceListUrl } = useUrl();
 
   const { t } = useTranslation();
 
   const { namespaceId, resourceName } = useParams();
-  const initialLayoutState = layout
-    ? {
-        layout: layout,
-        startColumn: {
-          resourceName: null,
-          resourceType: urlPath ?? resourceType,
-          namespaceId: namespaceId,
-          apiGroup: extension?.general.resource.group,
-          apiVersion: extension?.general.resource.version,
-        },
-        midColumn: {
-          resourceName: resourceName,
-          resourceType: urlPath ?? resourceType,
-          namespaceId: namespaceId,
-          apiGroup: extension?.general.resource.group,
-          apiVersion: extension?.general.resource.version,
-        },
-        endColumn: null,
-      }
-    : {
-        layout: layoutState?.layout,
-        startColumn: {
-          resourceType: urlPath ?? resourceType,
-          namespaceId: namespaceId,
-          apiGroup: extension?.general.resource.group,
-          apiVersion: extension?.general.resource.version,
-        },
-        midColumn: null,
-        endColumn: null,
-      };
 
-  useEffect(() => {
-    setLayoutColumn(initialLayoutState);
-    // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, [layout, namespaceId, resourceName, resourceType]);
+  usePrepareLayoutColumns({
+    resourceType: urlPath ?? resourceType,
+    namespaceId: namespaceId,
+    apiGroup: extension?.general.resource.group,
+    apiVersion: extension?.general.resource.version,
+    resourceName: resourceName,
+  });
 
   const overrides = { resourceType: urlPath };
 
@@ -83,7 +55,7 @@ const ColumnWrapper = ({
   );
 
   let startColumnComponent = null;
-  if (!layout && defaultColumn === 'details') {
+  if (layoutState.layout === 'OneColumn' && defaultColumn === 'details') {
     startColumnComponent = (
       <Details resourceName={resourceName} namespaceId={namespaceId} />
     );
