+++ src/shared/hooks/usePrepareLayout.js
@@ -1,3 +1,12 @@
+import { useEffect, useMemo } from 'react';
+import {
+  NavigationType,
+  useNavigationType,
+  useSearchParams,
+} from 'react-router-dom';
+import { useSetRecoilState } from 'recoil';
+import { columnLayoutState } from 'state/columnLayoutAtom';
+
 const switchToPrevLayout = layout => {
   switch (layout) {
     case 'EndColumn':
@@ -69,4 +78,74 @@ export function usePrepareLayout(layoutNumber) {
         ? `?layout=${nextLayout}`
         : '',
   };
+}
+
+export function usePrepareLayoutColumns({
+  resourceType,
+  namespaceId,
+  apiGroup,
+  apiVersion,
+  resourceName,
+}) {
+  const setLayoutColumn = useSetRecoilState(columnLayoutState);
+  const [searchParams] = useSearchParams();
+  const layout = searchParams.get('layout');
+  const showCreate = searchParams.get('showCreate');
+  const navigationType = useNavigationType();
+
+  const newLayoutState = useMemo(() => {
+    return layout
+      ? {
+          layout: layout,
+          startColumn: {
+            resourceType,
+            namespaceId,
+            apiGroup,
+            apiVersion,
+          },
+          midColumn:
+            !showCreate && resourceName
+              ? {
+                  resourceName,
+                  resourceType,
+                  namespaceId,
+                  apiGroup,
+                  apiVersion,
+                }
+              : null,
+          endColumn: null,
+          showCreate: showCreate ? { resourceType, namespaceId } : null,
+        }
+      : {
+          layout: 'OneColumn',
+          startColumn: {
+            resourceType,
+            namespaceId,
+            apiGroup,
+            apiVersion,
+          },
+          midColumn: null,
+          endColumn: null,
+          showCreate: null,
+        };
+  }, [
+    layout,
+    showCreate,
+    resourceType,
+    namespaceId,
+    apiGroup,
+    apiVersion,
+    resourceName,
+  ]);
+
+  useEffect(() => {
+    if (navigationType === NavigationType.Pop) {
+      setLayoutColumn(newLayoutState);
+    }
+  }, [newLayoutState, setLayoutColumn, navigationType]);
+
+  useEffect(() => {
+    setLayoutColumn(newLayoutState);
+    // eslint-disable-next-line react-hooks/exhaustive-deps
+  }, []);
 }
